+++ internal/provider/types/api_implementation_service.go
@@ -7,7 +7,6 @@ import (
 )
 
 type APIImplementationService struct {
-	AuthStrategySyncError *AuthStrategySyncError `tfsdk:"auth_strategy_sync_error"`
-	ControlPlaneID        types.String           `tfsdk:"control_plane_id"`
-	ID                    types.String           `tfsdk:"id"`
+	ControlPlaneID types.String `tfsdk:"control_plane_id"`
+	ID             types.String `tfsdk:"id"`
 }