+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 > Released on 2025/??/??
 
 ### BREAKING CHANGES
-* `auth_strategy_sync_error` read only field removed from `konnect_api`. It is now available on `konnect_api_implementation`
+* `auth_strategy_sync_error` read only field removed from `konnect_api`.
 
 ### Features
 * `current_version_summary` is now available on the `konnect_api` resource
+* `attributes` can now be set on the `konnect_api` resource
 
 ### Bug Fixes
 * `service` is now optional in `konnect_api_implementation`